refactor: move TUI-neutral runtime projections
