+++ core/crates/omegon/src/model_catalog.rs
